fix: reseed emptied workspaces after agent inventory settles
